Added option to align the ruler
I added the option to align the ruler left, right and center.

I also updated the spacer variable and ruler output code block. I'm looking for a way to eliminate the use of a spacer.gif image in order for the email to render pixel perfect in outlook.

Any suggestions on how to achieve perfect pixel height with empty cells in outlook?